feat: enhance UI with vertical gradient backgrounds across multiple screens, seperate homescreen into a component
