+++ app/src/main/java/com/personx/cryptx/screens/EncryptScreen.kt
@@ -3,6 +3,7 @@ package com.personx.cryptx.screens
 
 import android.content.ClipData
 import android.content.Context
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -22,6 +23,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ClipEntry
 import androidx.compose.ui.platform.LocalClipboard
@@ -63,6 +65,14 @@ fun EncryptScreen(
         modifier = Modifier
             .fillMaxSize()
             .padding(bottom = 16.dp)
+            .background(
+                Brush.verticalGradient(
+                    colors = listOf(
+                        MaterialTheme.colorScheme.onSurface.copy(0.05f),
+                        MaterialTheme.colorScheme.onPrimary.copy(0.01F)
+                    )
+                )
+            )
             .verticalScroll(rememberScrollState()),
         verticalArrangement = Arrangement.spacedBy(16.dp),
         horizontalAlignment = Alignment.CenterHorizontally
