+++ app/src/main/java/com/personx/cryptx/screens/HashGeneratorScreen.kt
@@ -22,6 +22,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.ClipEntry
 import androidx.compose.ui.platform.LocalClipboard
@@ -60,6 +61,14 @@ fun HashGeneratorScreen(
         modifier = Modifier
             .fillMaxSize()
             .padding(bottom = 16.dp)
+            .background(
+                Brush.verticalGradient(
+                    colors = listOf(
+                        MaterialTheme.colorScheme.onSurface.copy(0.05f),
+                        MaterialTheme.colorScheme.onPrimary.copy(0.01F)
+                    )
+                )
+            )
             .verticalScroll(rememberScrollState()),
         verticalArrangement = Arrangement.spacedBy(16.dp),
         horizontalAlignment = Alignment.CenterHorizontally
